rename and take out cookie func utility
